Isolate cursor.hpp
